+++ src/Adapter/Elasticsearch/Transaction.php
@@ -4,6 +4,10 @@ declare(strict_types = 1);
 namespace Formal\ORM\Adapter\Elasticsearch;
 
 use Formal\ORM\Adapter\Transaction as TransactionInterface;
+use Innmind\Immutable\{
+    Attempt,
+    SideEffect,
+};
 
 /**
  * @internal
@@ -23,33 +27,30 @@ final class Transaction implements TransactionInterface
     }
 
     #[\Override]
-    public function start(): void
+    public function start(): Attempt
     {
+        return Attempt::result(SideEffect::identity());
     }
 
     /**
      * @template R
      *
-     * @return callable(R): R
+     * @return callable(R): Attempt<R>
      */
     #[\Override]
     public function commit(): callable
     {
-        return static function(mixed $value) {
-            return $value;
-        };
+        return static fn(mixed $value) => Attempt::result($value);
     }
 
     /**
      * @template R
      *
-     * @return callable(R): R
+     * @return callable(R): Attempt<R>
      */
     #[\Override]
     public function rollback(): callable
     {
-        return static function(mixed $value) {
-            return $value;
-        };
+        return static fn(mixed $value) => Attempt::result($value);
     }
 }